alephscript/src/FIA/engine/smartpy/requirements.txt
click==7.1.2
dlib==19.20.0
Flask==1.1.2
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.5
onnx==1.7.0
onnxruntime==1.3.0
opencv-python==3.4.5.20
protobuf==3.12.2
PyYAML==5.3.1
six==1.15.0
typing-extensions==3.7.4.2
waitress==1.4.4
Werkzeug==1.0.1
